v1.2.5
Fix lighting bug (this happend because the game would restart the room's material to default)
Upgrade the procedure generator package to the newer version (edgar_2_0_0_alpha_4)
Fix enemy's shooting logic
Note: The AI of alien is not finished. This is a backup because i am about to refactor all the code to make it cleaner.